+++ Siv3D/src/Siv3D/Font/FontFace.cpp
@@ -114,10 +114,10 @@ namespace s3d
 		return GlyphInfo{
 			.glyphIndex	= glyphIndex,
 			.buffer		= 0,
-			.left		= static_cast<int32>(bbox.xMin),
-			.top		= static_cast<int32>(bbox.yMax),
-			.width		= static_cast<int32>(bbox.xMax - bbox.xMin),
-			.height		= static_cast<int32>(bbox.yMax - bbox.yMin),
+			.left		= static_cast<int16>(bbox.xMin),
+			.top		= static_cast<int16>(bbox.yMax),
+			.width		= static_cast<int16>(bbox.xMax - bbox.xMin),
+			.height		= static_cast<int16>(bbox.yMax - bbox.yMin),
 			.ascender	= m_property.ascender,
 			.descender	= m_property.descender,
 			.xAdvance	= (m_face->glyph->metrics.horiAdvance / 64.0),
@@ -137,10 +137,10 @@ namespace s3d
 
 		OutlineGlyph result;
 		result.glyphIndex	= glyphIndex;
-		result.left			= static_cast<int32>(bbox.xMin);
-		result.top			= static_cast<int32>(bbox.yMax);
-		result.width		= static_cast<int32>(bbox.xMax - bbox.xMin);
-		result.height		= static_cast<int32>(bbox.yMax - bbox.yMin);
+		result.left			= static_cast<int16>(bbox.xMin);
+		result.top			= static_cast<int16>(bbox.yMax);
+		result.width		= static_cast<int16>(bbox.xMax - bbox.xMin);
+		result.height		= static_cast<int16>(bbox.yMax - bbox.yMin);
 		result.ascender		= m_property.ascender;
 		result.descender	= m_property.descender;
 		result.xAdvance		= (m_face->glyph->metrics.horiAdvance / 64.0);
@@ -158,10 +158,10 @@ namespace s3d
 		SDFGlyph result;
 		result.glyphIndex	= glyphIndex;
 		result.buffer		= buffer;
-		result.left			= gi.left;
-		result.top			= gi.top;
-		result.width		= gi.width;
-		result.height		= gi.height;
+		result.left			= static_cast<int16>(gi.left);
+		result.top			= static_cast<int16>(gi.top);
+		result.width		= static_cast<int16>(gi.width);
+		result.height		= static_cast<int16>(gi.height);
 		result.ascender		= m_property.ascender;
 		result.descender	= m_property.descender;
 		result.xAdvance		= gi.xAdvance;
@@ -196,10 +196,10 @@ namespace s3d
 		MSDFGlyph result;
 		result.glyphIndex	= glyphIndex;
 		result.buffer		= buffer;
-		result.left			= static_cast<int32>(bbox.xMin);
-		result.top			= static_cast<int32>(bbox.yMax);
-		result.width		= width;
-		result.height		= height;
+		result.left			= static_cast<int16>(bbox.xMin);
+		result.top			= static_cast<int16>(bbox.yMax);
+		result.width		= static_cast<int16>(width);
+		result.height		= static_cast<int16>(height);
 		result.xAdvance		= (m_face->glyph->metrics.horiAdvance / 64.0);
 		result.yAdvance		= (m_face->glyph->metrics.vertAdvance / 64.0);
 		result.ascender		= m_property.ascender;
@@ -237,8 +237,8 @@ namespace s3d
 			m_property.styleName		= Unicode::Widen(m_face->style_name);
 			m_property.fontPixelSize	= pixelSize;
 			m_property.style			= style;
-			m_property.ascender			= static_cast<int32>(m_face->size->metrics.ascender / 64);
-			m_property.descender		= -static_cast<int32>(m_face->size->metrics.descender / 64);
+			m_property.ascender			= static_cast<int16>(m_face->size->metrics.ascender / 64);
+			m_property.descender		= -static_cast<int16>(m_face->size->metrics.descender / 64);
 		}
 
 		return true;
